test refactor with pytest-mock module and renaming
