+++ setup.py
@@ -22,8 +22,8 @@ install_requires = [
     ]
 
 tests_require = [
-    'mock',
     'pytest',
+    'pytest-mock',
     ]
 
 setup(
