+++ tests/exceptions_test.py
