+++ tests/sessionmixin_test.py
