+++ tox.ini
@@ -2,7 +2,7 @@
 envlist = py35,py27,py34
 [testenv]
 deps=pytest
-    mock
+    pytest-mock
     requests
     six
 commands=py.test  # or 'nosetests' or ...
